TooltipContent is made independent as a component and used in Tbas and FormElement
